requirements.txt
streamlit
google.generativeai
python-dotenv
mysql.connector
load_dotenv